SortingRecipePuzzleBasics
implemented a simple sorting puzzle. put "dropslots" where you want to place the object to be sorted, when the object is dropped on the correct "dropslot" it will snap into place and become non interactive. when the puzzle is solved the objects will fade out, and the recipe image will fade in. Currently all images are placeholders that need to be replaced, "dropslots" are also currently visible for testing, but they can easily be made to be invisible