Mortar Pestle Fail State 1
I made it so that if you pull the ingredient out of the mortar before its a powder there is a debug log saying it was too early
I also added the ingredients being destroyed if you add a second ingredient to the mortar (this can be changed to not destroy if that isnt what you want.
added an option to instantiate an explosion as well!