Sorting Puzzle, sprite fixes
I changed the boarder for the sprites to better reflect the image, for the more complicated shaped sprites we might have to actually change them in whatever drawing software they were made in.

made some updates to make the sorting a bit cleaner, made it so that the placed objects are in the background/picked up objects are in the foreground, and the drop slots overlapping is ignored allowing for more leeway in the sorting.